requirements.txt
requests[security] == 2.31.0
charset_normalizer == 3.3.2
chardet == 5.2.0
python-dotenv == 1.0.1